Remove INumber<T>
Remove usage of INumber<T> as it does not work with the .NET 6.0.101 SDK.

+++ src/AdventOfCode/Puzzles/Y2015/Day17.cs
@@ -28,7 +28,7 @@ public sealed class Day17 : Puzzle
     /// <returns>
     /// The combinations of containers that can store the volume specified by <paramref name="volume"/>.
     /// </returns>
-    internal static IList<ICollection<int>> GetContainerCombinations(int volume, IList<int> containerVolumes)
+    internal static IList<ICollection<long>> GetContainerCombinations(int volume, IList<int> containerVolumes)
     {
         var containers = containerVolumes
             .OrderByDescending((p) => p)
